+++ showcase/src/main/java/org/patternfly/showcase/component/NativePopoverComponent.java
@@ -20,10 +20,10 @@ import org.jboss.elemento.Elements;
 import org.jboss.elemento.router.Route;
 import org.patternfly.component.Severity;
 import org.patternfly.component.button.Button;
-import org.patternfly.component.popover.Popover2;
-import org.patternfly.component.popover.Popover2Body;
-import org.patternfly.component.popover.Popover2Footer;
-import org.patternfly.component.popover.Popover2Header;
+import org.patternfly.component.popover.NativePopover;
+import org.patternfly.component.popover.NativePopoverBody;
+import org.patternfly.component.popover.NativePopoverFooter;
+import org.patternfly.component.popover.NativePopoverHeader;
 import org.patternfly.component.togglegroup.ToggleGroup;
 import org.patternfly.showcase.LoremIpsum;
 import org.patternfly.showcase.Snippet;
@@ -38,8 +38,8 @@ import static org.patternfly.component.button.Button.button;
 import static org.patternfly.component.list.ActionList.actionList;
 import static org.patternfly.component.list.ActionListGroup.actionListGroup;
 import static org.patternfly.component.list.ActionListItem.actionListItem;
-import static org.patternfly.component.popover.Popover2.popover2;
-import static org.patternfly.component.popover.Popover2Body.popover2Body;
+import static org.patternfly.component.popover.NativePopover.nativePopover;
+import static org.patternfly.component.popover.NativePopoverBody.popoverBody;
 import static org.patternfly.component.togglegroup.ToggleGroup.toggleGroup;
 import static org.patternfly.component.togglegroup.ToggleGroupItem.toggleGroupItem;
 import static org.patternfly.icon.IconSets.fas.bullhorn;
@@ -55,67 +55,67 @@ import static org.patternfly.showcase.ApiDoc.Type.subcomponent;
 import static org.patternfly.showcase.Code.code;
 import static org.patternfly.showcase.Data.components;
 
-@Route(value = "/components/popover2", title = "Popover2")
-public class Popover2Component extends SnippetPage {
+@Route(value = "/components/native-popover", title = "Native Popover")
+public class NativePopoverComponent extends SnippetPage {
 
-    public Popover2Component() {
-        super(components.get("popover2"));
+    public NativePopoverComponent() {
+        super(components.get("native-tooltip"));
 
         startExamples();
-        addSnippet(new Snippet("popover2-basic", "Basic",
-                code("popover2-basic"), () ->
-                // @code-start:popover2-basic
+        addSnippet(new Snippet("np-basic", "Basic",
+                code("np-basic"), () ->
+                // @code-start:np-basic
                 div().style("margin", "50px")
                         .add(button("Toggle popover").primary()
-                                .id("popover2-basic-button"))
-                        .add(popover2(By.id("popover2-basic-button"))
+                                .id("np-basic-button"))
+                        .add(nativePopover(By.id("np-basic-button"))
                                 .ariaLabel("Basic popover")
                                 .addHeader("Popover header")
                                 .addBody("Popovers are triggered by click rather than hover.")
                                 .addFooter("Popover footer"))
                         .element()
-                // @code-end:popover2-basic
+                // @code-end:np-basic
         ));
 
-        addSnippet(new Snippet("popover2-placements", "Placements",
-                code("popover2-placements"), () ->
-                // @code-start:popover2-placements
+        addSnippet(new Snippet("np-placements", "Placements",
+                code("np-placements"), () ->
+                // @code-start:np-placements
                 div().style("display: flex; gap: 16px")
                         .add(button("Bottom").secondary()
-                                .id("popover2-bottom-button"))
-                        .add(popover2(By.id("popover2-bottom-button"))
+                                .id("nt-placements-bottom-button"))
+                        .add(nativePopover(By.id("nt-placements-bottom-button"))
                                 .placement(bottom)
                                 .ariaLabel("Bottom popover")
                                 .addHeader("Bottom popover")
                                 .addBody("This popover is placed at the bottom.")
                                 .addFooter("Popover footer"))
                         .add(button("Left").secondary()
-                                .id("popover2-left-button"))
-                        .add(popover2(By.id("popover2-left-button"))
+                                .id("nt-placements-left-button"))
+                        .add(nativePopover(By.id("nt-placements-left-button"))
                                 .placement(left)
                                 .ariaLabel("Left popover")
                                 .addHeader("Left popover")
                                 .addBody("This popover is placed on the left.")
                                 .addFooter("Popover footer"))
                         .add(button("Right").secondary()
-                                .id("popover2-right-button"))
-                        .add(popover2(By.id("popover2-right-button"))
+                                .id("nt-placements-right-button"))
+                        .add(nativePopover(By.id("nt-placements-right-button"))
                                 .placement(right)
                                 .ariaLabel("Right popover")
                                 .addHeader("Right popover")
                                 .addBody("This popover is placed on the right.")
                                 .addFooter("Popover footer"))
                         .element()
-                // @code-end:popover2-placements
+                // @code-end:np-placements
         ));
 
-        addSnippet(new Snippet("popover2-close", "Close popover",
-                code("popover2-close"), () -> {
-            // @code-start:popover2-close
-            Popover2 popover = popover2(By.id("popover2-close-button"));
+        addSnippet(new Snippet("np-close", "Close popover",
+                code("np-close"), () -> {
+            // @code-start:np-close
+            NativePopover popover = nativePopover(By.id("np-close-button"));
             popover.ariaLabel("Popover with button in the body that can close it")
                     .addHeader("Popover header")
-                    .addBody(popover2Body()
+                    .addBody(popoverBody()
                             .add(div()
                                     .add(div()
                                             .add("A controlled popover can be closed by calling its ")
@@ -128,46 +128,46 @@ public class Popover2Component extends SnippetPage {
 
             return div().style("margin", "50px")
                     .add(button("Toggle popover").primary()
-                            .id("popover2-close-button"))
+                            .id("np-close-button"))
                     .add(popover)
                     .element();
-            // @code-end:popover2-close
+            // @code-end:np-close
         }));
 
-        addSnippet(new Snippet("popover2-no-padding", "Without header/footer/close and no padding",
-                code("popover2-no-padding"), () ->
-                // @code-start:popover2-no-padding
+        addSnippet(new Snippet("np-no-padding", "Without header/footer/close and no padding",
+                code("np-no-padding"), () ->
+                // @code-start:np-no-padding
                 div().style("margin", "50px")
                         .add(button("Toggle popover").primary()
-                                .id("popover2-no-padding-button"))
-                        .add(popover2(By.id("popover2-no-padding-button"))
+                                .id("np-no-padding-button"))
+                        .add(nativePopover(By.id("np-no-padding-button"))
                                 .noClose()
                                 .noPadding()
                                 .ariaLabel("Popover with no header, footer, close button, and padding")
                                 .addBody(
                                         "This popover has no padding and is intended for use with content that has its own spacing and should touch the edges of the popover container."))
                         .element()
-                // @code-end:popover2-no-padding
+                // @code-end:np-no-padding
         ));
 
-        addSnippet(new Snippet("popover2-auto-width", "Width auto",
-                code("popover2-auto-width"), () ->
-                // @code-start:popover2-auto-width
+        addSnippet(new Snippet("np-auto-width", "Width auto",
+                code("np-auto-width"), () ->
+                // @code-start:np-auto-width
                 div().style("margin", "50px")
                         .add(button("Toggle popover").primary()
-                                .id("popover2-auto-width-button"))
-                        .add(popover2(By.id("popover2-auto-width-button"))
+                                .id("np-auto-width-button"))
+                        .add(nativePopover(By.id("np-auto-width-button"))
                                 .autoWidth()
                                 .ariaLabel("Popover with auto-width.")
                                 .addBody("Removes fixed-width and allows width to be defined by contents."))
                         .element()
-                // @code-end:popover2-auto-width
+                // @code-end:np-auto-width
         ));
 
-        addSnippet(new Snippet("popover2-auto-placement", "Auto placement",
+        addSnippet(new Snippet("np-auto-placement", "Auto placement",
                 "You might need to resize the browser window to a minimum to see the auto placement in action.",
-                code("popover2-auto-placement"), () -> {
-            // @code-start:popover2-auto-placement
+                code("np-auto-placement"), () -> {
+            // @code-start:np-auto-placement
             Button button = button("Toggle popover");
             onAttach(button, mr -> {
                 ScrollIntoViewOptions options = ScrollIntoViewOptions.create();
@@ -176,7 +176,7 @@ public class Popover2Component extends SnippetPage {
                 button.element().scrollIntoView(options);
             });
 
-            Popover2 popover = popover2(By.id("popover2-auto-placement-button"));
+            NativePopover popover = nativePopover(By.id("np-auto-placement-button"));
             ToggleGroup toggleGroup = toggleGroup(single)
                     .addItem(toggleGroupItem("left", "left").onClick((e, t) -> popover.placement(left)))
                     .addItem(toggleGroupItem("top", "top").onClick((e, t) -> popover.placement(top)))
@@ -191,7 +191,7 @@ public class Popover2Component extends SnippetPage {
                             .add(div().style("width", "720px")
                                     .add(div().css("pfj-popover-box")
                                             .add(button.css("pfj-popover-button").primary()
-                                                    .id("popover2-auto-placement-button"))
+                                                    .id("np-auto-placement-button"))
                                             .add(popover
                                                     .placement(auto)
                                                     .ariaLabel("Advanced popover")
@@ -199,74 +199,74 @@ public class Popover2Component extends SnippetPage {
                                                     .addBody(LoremIpsum.words(20))
                                                     .addFooter("Popover footer")))))
                     .element();
-            // @code-end:popover2-auto-placement
+            // @code-end:np-auto-placement
         }));
 
-        addSnippet(new Snippet("popover2-icon", "Popover with icon in the title",
-                code("popover2-icon"), () ->
-                // @code-start:popover2-icon
+        addSnippet(new Snippet("np-icon", "Popover with icon in the title",
+                code("np-icon"), () ->
+                // @code-start:np-icon
                 div().style("margin", "50px")
                         .add(button("Toggle popover").primary()
-                                .id("popover2-icon-button"))
-                        .add(popover2(By.id("popover2-icon-button"))
+                                .id("np-icon-button"))
+                        .add(nativePopover(By.id("np-icon-button"))
                                 .icon(bullhorn())
                                 .ariaLabel("Popover with icon in the title example.")
                                 .addHeader("Popover with icon")
                                 .addBody(LoremIpsum.words(20))
                                 .addFooter("Popover footer"))
                         .element()
-                // @code-end:popover2-icon
+                // @code-end:np-icon
         ));
 
-        addSnippet(new Snippet("popover2-alert", "Alert popover",
-                code("popover2-alert"), () -> {
-            // @code-start:popover2-alert
+        addSnippet(new Snippet("np-alert", "Alert popover",
+                code("np-alert"), () -> {
+            // @code-start:np-alert
             return div().style("margin", "50px")
                     .add(actionList()
                             .addItem(actionListGroup()
                                     .addItem(actionListItem()
-                                            .add(button("Custom").secondary().id("popover2-alert-custom")))
+                                            .add(button("Custom").secondary().id("np-alert-custom")))
                                     .addItem(actionListItem()
-                                            .add(button("Info").secondary().id("popover2-alert-info")))
+                                            .add(button("Info").secondary().id("np-alert-info")))
                                     .addItem(actionListItem()
-                                            .add(button("Success").secondary().id("popover2-alert-success")))
+                                            .add(button("Success").secondary().id("np-alert-success")))
                                     .addItem(actionListItem()
-                                            .add(button("Warning").warning().id("popover2-alert-warning")))
+                                            .add(button("Warning").warning().id("np-alert-warning")))
                                     .addItem(actionListItem()
-                                            .add(button("Danger").danger().id("popover2-alert-danger")))))
-                    .add(popover2(By.id("popover2-alert-custom"))
+                                            .add(button("Danger").danger().id("np-alert-danger")))))
+                    .add(nativePopover(By.id("np-alert-custom"))
                             .status(Severity.custom)
                             .addHeader("Popover custom header")
                             .addBody(LoremIpsum.words(20))
                             .addFooter("Popover footer"))
-                    .add(popover2(By.id("popover2-alert-info"))
+                    .add(nativePopover(By.id("np-alert-info"))
                             .status(Severity.info)
                             .addHeader("Popover info header")
                             .addBody(LoremIpsum.words(20))
                             .addFooter("Popover footer"))
-                    .add(popover2(By.id("popover2-alert-success"))
+                    .add(nativePopover(By.id("np-alert-success"))
                             .status(Severity.success)
                             .addHeader("Popover success header")
                             .addBody(LoremIpsum.words(20))
                             .addFooter("Popover footer"))
-                    .add(popover2(By.id("popover2-alert-warning"))
+                    .add(nativePopover(By.id("np-alert-warning"))
                             .status(Severity.warning)
                             .addHeader("Popover warning header")
                             .addBody(LoremIpsum.words(20))
                             .addFooter("Popover footer"))
-                    .add(popover2(By.id("popover2-alert-danger"))
+                    .add(nativePopover(By.id("np-alert-danger"))
                             .status(Severity.danger)
                             .addHeader("Popover danger header")
                             .addBody(LoremIpsum.words(20))
                             .addFooter("Popover footer"))
                     .element();
-            // @code-end:popover2-alert
+            // @code-end:np-alert
         }));
 
-        startApiDocs(Popover2.class);
-        addApiDoc(Popover2.class, component);
-        addApiDoc(Popover2Body.class, subcomponent);
-        addApiDoc(Popover2Footer.class, subcomponent);
-        addApiDoc(Popover2Header.class, subcomponent);
+        startApiDocs(NativePopover.class);
+        addApiDoc(NativePopover.class, component);
+        addApiDoc(NativePopoverBody.class, subcomponent);
+        addApiDoc(NativePopoverFooter.class, subcomponent);
+        addApiDoc(NativePopoverHeader.class, subcomponent);
     }
 }